No auto overwrite config files

+++ src/setup.js
@@ -14,21 +14,23 @@ const serverPort = argv.serverPort || 21025;
 const relayPort = argv.relayPort || 2003;
 
 function UpdateEnvFile() {
+    const envFile = join(__dirname, '../.env');
+    if (fs.existsSync(envFile) && !argv.force) return console.log('Env file already exists, use --force to overwrite it');
+    
     const exampleEnvFilePath = join(__dirname, '../example.env');
-
     let exampleEnvText = fs.readFileSync(exampleEnvFilePath, 'utf8');
     exampleEnvText = exampleEnvText.replaceAll('{{ grafanaPort }}', grafanaPort).replaceAll('{{ serverPort }}', serverPort);
-    const envFile = join(__dirname, '../.env');
     fs.writeFileSync(envFile, exampleEnvText);
     console.log('Env file created');
 }
 
 async function UpdateDockerComposeFile() {
-    const exampleDockerComposeFile = join(__dirname, '../docker-compose.example.yml');
+    const dockerComposeFile = join(__dirname, '../docker-compose.yml');
+    if (fs.existsSync(dockerComposeFile) && !argv.force) return console.log('Docker-compose file already exists, use --force to overwrite it');
 
+    const exampleDockerComposeFile = join(__dirname, '../docker-compose.example.yml');
     let exampleDockerComposeText = fs.readFileSync(exampleDockerComposeFile, 'utf8');
     exampleDockerComposeText = exampleDockerComposeText.replaceAll('{{ grafanaPort }}', grafanaPort).replaceAll('{{ serverPort }}', serverPort).replaceAll('{{ relayPort }}', relayPort);
-    const dockerComposeFile = join(__dirname, '../docker-compose.yml');
     fs.writeFileSync(dockerComposeFile, exampleDockerComposeText);
     console.log('Docker-compose file created');
 }
